requirements.txt
pandas
yfinance
feedparser
google-generativeai
python-dotenv
requests
schedule
pytz
